backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
transformers
torch
pdfplumber
python-docx
textblob
spacy